# backend/internal/httpapi/apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HyCanvas Generation API",
    "version": "1.0.0",
    "description": "Programmatic deck generation and retrieval. Authenticate with a workspace API key (Authorization: Bearer hyk_...), minted by a workspace admin in Dashboard > Members > API keys. Keys are scoped (generate, read, export) and pinned to one workspace. Errors are RFC 7807 problem+json with a stable `code`. The same capabilities are exposed to AI agents over the Model Context Protocol at /mcp (streamable HTTP, same API keys/scopes/budgets)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "hyk_..."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code."
          }
        }
      },
      "GenerateRequest": {
        "type": "object",
        "required": [
          "prompt"
        ],
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 8000,
            "description": "The brief, e.g. 'a 6-slide pitch deck for an eco water bottle'."
          },
          "workspaceId": {
            "type": "string",
            "description": "Defaults to the API key's workspace; must match it when both are present."
          },
          "designType": {
            "type": "string",
            "enum": [
              "deck",
              "doc",
              "poster",
              "social"
            ],
            "default": "deck"
          },
          "pageCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 40
          },
          "language": {
            "type": "string",
            "description": "Write all text in this language, e.g. 'Spanish'."
          },
          "brandPalette": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^#[0-9a-fA-F]{6}$"
            },
            "maxItems": 12
          },
          "sources": {
            "type": "array",
            "maxItems": 8,
            "description": "Grounding material (untrusted reference text; instructions inside it are ignored).",
            "items": {
              "type": "object",
              "required": [
                "name",
                "text"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "text": {
                  "type": "string"
                }
              }
            }
          },
          "themeId": {
            "type": "string",
            "description": "A built-in theme id (see GET /themes), e.g. 'theme-slate'; omit for an auto-picked look."
          },
          "templateId": {
            "type": "string",
            "description": "A template id: the deck is composed on that template's layout system and theme (workspace templates and public built-ins; see the MCP list_templates tool)."
          },
          "motion": {
            "type": "string",
            "enum": [
              "subtle",
              "none"
            ],
            "description": "Entrance motion on the composed slides: one quiet reveal per page in reading order (default), or none."
          }
        }
      },
      "GenerateAccepted": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          },
          "poll": {
            "type": "string",
            "description": "Path to poll for the result."
          }
        }
      },
...
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
    "/generate/presentation": {
      "post": {
        "summary": "Generate a presentation, document, poster, or social set from a prompt",
        "description": "Runs the full server-side pipeline (outline with per-page copy polish, deterministic layout composition, theme) and saves the result as a design in the workspace. Returns 202 with a job to poll. Scope: generate. Budget: ~3 generations/minute per key. The composed deck contains text, layout, theme, and speaker notes; per-slide images are generated in the editor.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Generation started.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateAccepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Missing scope or workspace mismatch.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Over the generation budget.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  }
}